Numeric.LinearAlgebra.Backprop

Documentation

dot :: (Numeric t, Backprop t, Reifies s W) => BVar s (Vector t) -> BVar s (Vector t) -> BVar s t

(<.>) :: (Numeric t, Backprop t, Reifies s W) => BVar s (Vector t) -> BVar s (Vector t) -> BVar s t infixr 8

Infix synonym for dot

(#>) :: (Reifies s W, Product t, Numeric t, Backprop t) => BVar s (Matrix t) -> BVar s (Vector t) -> BVar s (Vector t) infixr 8

(<#) :: (Reifies s W, Product t, Numeric t, Backprop t) => BVar s (Vector t) -> BVar s (Matrix t) -> BVar s (Vector t) infixr 8

(<>) :: (Reifies s W, Product t, Numeric t) => BVar s (Matrix t) -> BVar s (Matrix t) -> BVar s (Matrix t) infixr 8

outer :: (Reifies s W, Product t, Numeric t, Backprop t) => BVar s (Vector t) -> BVar s (Vector t) -> BVar s (Matrix t)

kronecker :: (Reifies s W, Product t, Numeric t) => BVar s (Matrix t) -> BVar s (Matrix t) -> BVar s (Matrix t)

cross :: (Product t, Backprop t, Reifies s W) => BVar s (Vector t) -> BVar s (Vector t) -> BVar s (Vector t)

Only defined for inputs that are three-element vectors

sumElements :: (Backprop e, Backprop (c e), Container c e, Konst e (IndexOf c) c, Reifies s W) => BVar s (c e) -> BVar s e

flatten :: (Element t, Container Vector t, Num t, Backprop t, Reifies s W) => BVar s (Matrix t) -> BVar s (Vector t)

Orphan instances

(Container Vector a, Num a) => Backprop (Matrix a)

Methods

zero :: Matrix a -> Matrix a

add :: Matrix a -> Matrix a -> Matrix a

one :: Matrix a -> Matrix a